# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for a delivery act upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [deliveryId, name, size, contentType]
      properties:
        deliveryId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
...
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
          description: Normalized object path (e.g. /objects/uploads/uuid). Store this in the database.